OpenCL: change all doubles to floats for cross-OS compatibility; fix local work size exceeding max limits

+++ solver.h
@@ -8,7 +8,7 @@
 typedef struct{
     int *row_ptr;
     int *col_ind;
-    double *values;
+    float *values;
 } CSRMatrix;
 
 typedef struct{
@@ -27,25 +27,25 @@ typedef struct{
 typedef struct {
     int width, height;
     int time_step;  //Curent simulation index (starts from 0)
-    double dx, dy;  //Physical distance 
-    double dt;      //Physical time between steps
-    double rx, ry;  //Numerical diffusion coefficients 
-    double alpha;   //Diffusion coefficient
+    float dx, dy;  //Physical distance 
+    float dt;      //Physical time between steps
+    float rx, ry;  //Numerical diffusion coefficients 
+    float alpha;   //Diffusion coefficient
 
     float *u_current;  //Current state of the system
     float *u_next;     //Next state of the system
     CSRMatrix A;  //Sparse matrix in CSR format for the system of equations
-    double *b;    //Right-hand side vector for the system of equations
+    float *b;    //Right-hand side vector for the system of equations
 
     OpenCLContext cl;
 } Solver;
 
-Solver *setup_solver(int width, int height, double dx, double dy, double dt, double alpha, float *u_curr);
+Solver *setup_solver(int width, int height, float dx, float dy, float dt, float alpha, float *u_curr);
 void update_system(Solver *solver);
 CSRMatrix allocate_CSR_matrix(int width, int height);
 void free_solver(Solver *solver);
 void free_CSR_matrix(CSRMatrix *matrix);
-void setup_coefficients_matrix(double rx, CSRMatrix *A, int width, int height);
+void setup_coefficients_matrix(float rx, CSRMatrix *A, int width, int height);
 void debug_print_CSR(const CSRMatrix *A, int n);
 void setup_opencl_context(Solver *solver);
 cl_event populate_b(Solver *solver);
